<refactor>[frontend]: Centralizado rotas em routes
